moved upserts to activities
